Handle recurring transactions and company switching on Projections page
